fix(kit): resolve circular dependencies inside `Number` mask
