+++ projects/kit/src/lib/constants/default-pseudo-minuses.ts
@@ -1,0 +1,15 @@
+import {
+    CHAR_EM_DASH,
+    CHAR_EN_DASH,
+    CHAR_HYPHEN,
+    CHAR_JP_HYPHEN,
+    CHAR_MINUS,
+} from './unicode-characters';
+
+export const DEFAULT_PSEUDO_MINUSES = [
+    CHAR_HYPHEN,
+    CHAR_EN_DASH,
+    CHAR_EM_DASH,
+    CHAR_JP_HYPHEN,
+    CHAR_MINUS,
+];
