+++ projects/kit/src/lib/constants/index.ts
@@ -1,6 +1,7 @@
 export * from './date-segment-max-values';
 export * from './default-decimal-pseudo-separators';
 export * from './default-min-max-dates';
+export * from './default-pseudo-minuses';
 export * from './default-time-segment-bounds';
 export * from './meridiem';
 export * from './time-fixed-characters';
