+++ projects/kit/src/lib/masks/number/index.ts
@@ -3,3 +3,9 @@ export * from './number-params';
 export * from './plugins';
 export * from './processors';
 export * from './utils';
+/*
+Don't put it inside `./utils` entrypoint to avoid circular dependency
+- `stringify-number.ts` uses `maskitoNumberOptionsGenerator`
+- almost all processors and plugins (which are part of `maskitoNumberOptionsGenerator`) uses `./utils`
+ */
+export * from './utils/stringify-number';
