+++ projects/kit/src/lib/masks/number/utils/index.ts
@@ -2,6 +2,5 @@ export * from './extract-affixes';
 export * from './generate-mask-expression';
 export * from './number-parts';
 export * from './parse-number';
-export * from './stringify-number';
 export * from './stringify-number-without-exp';
 export * from './validate-decimal-pseudo-separators';
